requirements.txt
flask
numpy
pandas
joblib
scikit-learn
pyserial